UX-396# durationInMillis added to node details object.
